+++ cypress/tests/guillotina/sharing.js
@@ -1,4 +1,4 @@
-describe('History Tests', () => {
+describe('Sharing Tests', () => {
   beforeEach(() => {
     // give a logged in editor and the site root
     cy.autologin();
@@ -9,7 +9,7 @@ describe('History Tests', () => {
     cy.waitForResourceToLoad('@types');
     cy.waitForResourceToLoad('');
   });
-  it('As editor I can add a page and access history', function () {
+  it('As editor I can add a page and access sharing', function () {
     // I add a page
     cy.get('#toolbar-add').click();
     cy.get('#toolbar-add-document').click();
@@ -21,16 +21,15 @@ describe('History Tests', () => {
     // then a new page has been created
     cy.get('#toolbar-save').click();
     cy.url().should('eq', Cypress.config().baseUrl + '/my-page');
-
-    cy.get('.navigation .item.active').should('have.text', 'My Page');
+    cy.contains('My Page');
 
     // then I click on the Toolbar > More
     cy.get('#toolbar-more').click();
-    cy.get('.menu-more').contains('History');
+    cy.get('.menu-more').contains('Sharing');
 
     // and then I click on History
-    cy.get('.menu-more a[href*="/history"]').contains('History').click();
-    cy.url().should('eq', Cypress.config().baseUrl + '/my-page/history');
-    cy.contains('History of');
+    cy.get('.menu-more a[href*="/sharing"]').contains('Sharing').click();
+    cy.url().should('eq', Cypress.config().baseUrl + '/my-page/sharing');
+    cy.contains('Sharing for');
   });
 });